+++ src/cpu/control_register.js
@@ -158,6 +158,18 @@ class ControlRegister extends CpuRegister {
     ifhalfcarry = () => {
         return (this.value & cpus.HALFCARRY) > 0;
     }
+
+    ifentireset = () => {
+        return (this.value & cpus.ENTIRE) > 0;
+    }
+
+    ifirqclear = () => {
+        return (this.value & cpus.IRQ) === 0;
+    }
+
+    iffirqclear = () => {
+        return (this.value & cpus.FIRQ) === 0
+    }
 }
 
 module.exports = { ControlRegister }