+++ src/cpu/instructions.js
@@ -1,3 +1,5 @@
+const cpus = require('../../src/cpu/cpu_constants');
+
 instructions = {
   0x00: {
     'operation': 'NEG',
@@ -203,6 +205,18 @@ instructions = {
       'READHIGH', 'READWLOW', 'ADDTGSWTOOBIF',
     ],
   },
+  0x103f: {
+    'operation': 'SWI2',
+    'mode': 'inherent',
+    'object': 'S',
+    'target': 'AD',
+    'vector': cpus.vSWI2.vector,
+    'mask': 0x00,
+    'code': [
+      'SETENTIRE', 'TFROBTOTG', 'PUSH', 'TFRTGTOOB',
+      'VECTORHIGH', 'VECTORLOW', 'MASKIF',
+    ],
+  },
   0x108c: {
     'operation': 'CMPY',
     'mode': 'immediate',
@@ -322,6 +336,18 @@ instructions = {
     'operation': 'extended instruction',
     'mode': 'fetch',
   },
+  0x113f: {
+    'operation': 'SWI3',
+    'mode': 'inherent',
+    'object': 'S',
+    'target': 'AD',
+    'vector': cpus.vSWI3.vector,
+    'mask': 0,
+    'code': [
+      'SETENTIRE', 'TFROBTOTG', 'PUSH', 'TFRTGTOOB',
+      'VECTORHIGH', 'VECTORLOW', 'MASKIF',
+    ],
+  },
   0x1183: {
     'operation': 'CMPU',
     'mode': 'immediate',
@@ -378,9 +404,11 @@ instructions = {
   },
   0x12: {
     'operation': 'NOP',
-    'code': [
-      'BUSY',
-    ],
+    'code': ['BUSY'],
+  },
+  0x13: {
+    'operation': 'SYNC',
+    'code': ['SYNC'],
   },
   0x16: {
     'operation': 'LBRA',
@@ -636,8 +664,28 @@ instructions = {
     'operation': 'RTS',
     'mode': 'immediate',
     'object': 'S',
+    'target': 'AD',
     'code': [
-      'PULLPC', 'BUSY', 'BUSY', 'BUSY',
+      'TFROBTOTG', 'PULLPC', 'TFRTGTOOB', 'BUSY',
+    ],
+  },
+  0x3b: {
+    'operation': 'RTI',
+    'mode': 'inherent',
+    'object': 'S',
+    'target': 'AD',
+    'code': [
+      'TFROBTOTG', 'PULLCC', 'PULLPC', 'TFRTGTOOB', 'BUSY',
+    ],
+  },
+  0x3c: {
+    'operation': 'CWAI',
+    'mode': 'immediate',
+    'object': 'S',
+    'target': 'AD',
+    'code': [
+      'READWLOW', 'ANDCC', 'SETENTIRE', 'TFROBTOTG', 'PUSH',
+      'TFRTGTOOB', 'BUSY', 'WAIT',
     ],
   },
   0x3d: {
@@ -646,6 +694,18 @@ instructions = {
     'code': [
       'MULTIPLY', 'BUSY', 'BUSY', 'BUSY', 'BUSY',
       'BUSY', 'BUSY', 'BUSY', 'BUSY', 'BUSY',
+    ],
+  },
+  0x3f: {
+    'operation': 'SWI',
+    'mode': 'inherent',
+    'object': 'S',
+    'target': 'AD',
+    'vector': cpus.vSWI.vector,
+    'mask': cpus.IRQ | cpus.FIRQ,
+    'code': [
+      'SETENTIRE', 'TFROBTOTG', 'PUSH', 'TFRTGTOOB',
+      'VECTORHIGH', 'VECTORLOW', 'MASKIF',
     ],
   },
   0x40: {
@@ -1685,6 +1745,37 @@ instructions = {
     'target': 'AD',
     'code': [
       'READHIGH', 'READWLOW', 'TFRWTOTG', 'WRITEHIGH', 'WRITELOW',
+    ],
+  },
+  'fastVectorFromRun': {
+    'operation': 'HWI',
+    'mode': 'inherent',
+    'object': 'S',
+    'target': 'AD',
+    'code': [
+      'TFROBTOTG', 'PUSHIR', 'TFRTGTOOB',
+      'VECTORHIGH', 'VECTORLOW', 'MASKIF',
+      'BUSY', 'BUSY', 'BUSY',
+    ],
+  },
+  'vectorFromRun': {
+    'operation': 'HWI',
+    'mode': 'inherent',
+    'object': 'S',
+    'target': 'AD',
+    'code': [
+      'SETENTIRE', 'TFROBTOTG', 'PUSH', 'TFRTGTOOB',
+      'VECTORHIGH', 'VECTORLOW', 'MASKIF', 'BUSY', 'BUSY',
+    ],
+  },
+  'vectorFromWait': {
+    'operation': 'HWI',
+    'mode': 'inherent',
+    'object': 'S',
+    'target': 'AD',
+    'code': [
+      'VECTORHIGH', 'VECTORLOW', 'MASKIF',
+      'BUSY', 'BUSY', 'BUSY', 'BUSY', 'BUSY',
     ],
   },
 };